Restores exception throwing if board not found
Will be compatible with liboncilla lazy initialization of nodes

Was a mistake as the entry in the intern table is always here, but it
could be storing a NULL shared_ptr.

I think the removing of exception throwing has been done to use the
robot with less than 4 leg. People should not rely on this trick to do
it, but rather use
rci::oncilla::Oncilla::throwExceptionOnInitialization(false), to
disable exception throwing if any of the leg is not found. Then the
exception will be defered to the actual getLX() call that want to use
a unavailable leg.